Fixed a few bugs in DNAProperty, and added the models file to version control.
Since we have three different model types and creating separate CSVs for them would be an enormous pain, we can't use the bulk loader. fasta2datastore.py is a dirty hack that will create a datastore locally; that datastore can then be exported and uploaded to appspot. It's not quite finished yet, and you really shouldn't do this -- I shouldn't be doing it either, but I'm a bad person.